LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/starting-style-name-defining-rules-expected.txt

XXXX

FAIL @keyframes and @layer in @starting-style apply assert_equals: @keyframes applied expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
FAIL Load @font-face from @starting-style rule assert_equals: expected "80px" but got "57.78125px"
